+++ packages/terra-aggregator/src/terra-dev-site/doc/common/placeholder-list/PlaceholderListItem.jsx
@@ -1,0 +1,34 @@
+import React from 'react';
+import './PlaceholderList.scss';
+
+/* eslint-disable */
+const PlaceholderListItem = ({
+  children,
+  isSelected,
+  onSelect,
+  ...customProps
+}) => {
+  const keyDown = event => {
+    if (event.nativeEvent.keyCode === 13 || event.nativeEvent.keyCode === 32) {
+      event.preventDefault();
+      onSelect(event);
+    }
+  };
+
+  const classNames = isSelected ? 'placeholder-list-item is-selected' : 'placeholder-list-item ';
+  return (
+    <li
+      {...customProps} 
+      aria-selected={isSelected}
+      tabIndex="0"
+      className={classNames}
+      onClick={onSelect}
+      onKeyDown={keyDown}
+      role="option"
+    >
+      {children}
+    </li>
+  );
+};
+
+export default PlaceholderListItem;