+++ packages/terra-application-links/CHANGELOG.md
@@ -3,8 +3,16 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added en-AU translations.
+
 ### Changed
 * Update external links example to point to terra-ui.com
+
+5.2.0 - (February 13, 2019)
+------------------
+### Changed
+* Removed dependency on terra-list
 
 5.1.0 - (February 5, 2019)
 ------------------