+++ packages/terra-collapsible-menu-view/CHANGELOG.md
@@ -3,6 +3,16 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+* Added en-AU translations
+
+5.0.0 - (February 13, 2019)
+------------------
+### Breaking Change
+* Major dependency version bump of terra-list
+
+### Changed
+* Wrapped Collapsible Menu Items in <li>
 
 4.3.0 - (February 5, 2019)
 ------------------