+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -3,8 +3,16 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+* Added en-AU translations
+
 ### Fixed
 * Changed browser.execute functions in wdio spec to be strings for IE and FF driver compatibility
+
+3.6.0 - (February 13, 2019)
+------------------
+### Fixed
+* Throws NPE if manually entered valid date is outside of allotted datetime range.
 
 3.5.0 - (February 5, 2019)
 ------------------