+++ packages/terra-time-input/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added en-AU translations
+
 ### Fixed
 * Changed browser.execute function in wdio spec to be a string for IE and FF driver compatibility